requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
motor==3.6.0
python-docx==1.1.2
pypdf==5.1.0
python-multipart==0.0.12
telethon==1.36.0
